First commit: add core files
Added engine.h, which contains the "game engine" functions, example.c uses these functions in an example "game". I'm not a great game designer, so the example may not be very fun or even a game. Currently, 4 buttons are shown on the screen that show their corresponding cards when held.